+++ packages/components/organisms/f-header/test/visual/f-header.visual.mobile.spec.js
@@ -53,6 +53,30 @@ describe('Shared - f-header component tests', () => {
         browser.percyScreenshot(`f-header - Theme colours - ${theme}`, 'mobile');
     });
 
+    forEach([
+        'white',
+        'highlight',
+        'transparent'
+    ]).it('should display the mobile nav correctly for theme %s', theme => {
+        // Arrange
+        const controls = [
+            'locale:en-GB',
+            'showOffersLink:true',
+            'showDeliveryEnquiry:true',
+            `headerBackgroundTheme:${theme}`
+        ].join(';');
+
+        header = new Header();
+        header.path += `&args=${controls}`;
+
+        // Act
+        header.load();
+        header.openMobileNavigationBar();
+
+        // Assert
+        browser.percyScreenshot(`f-header - Mobile nav theme colours - ${theme}`, 'mobile');
+    });
+
     it('should display all available countries', () => {
         // Arrange
         const controls = 'locale:en-GB';